pyproject.toml
[tool.poetry]
name = "fractal-database-matrix"
version = "0.0.7"
description = ""
authors = ["Justin <justin@fractalnetworks.co"]
readme = "README.md"
include = ["synapse/**"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.0"
matrix-nio = "^0.22.1"
fractal-matrix-client = ">=0.0.1"
taskiq-matrix = ">=0.0.1"
fractal-cli = ">=0.0.1"
pytest = { version = "^7.4.3", optional = true }
pytest-asyncio = { version = "^0.21.1", optional = true }
pytest-cov = { version = "^4.1.0", optional = true }
pytest-mock = { version = "^3.11.1", optional = true }
ipython = { version = "^8.17.2", optional = true }
pytest-django = { version = "^4.5.2", optional = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
dev = ["pytest-django", "pytest", "pytest-cov", "pytest-mock", "pytest-asyncio", "ipython"]
